+++ tests/integration/snapshot.rs
@@ -332,6 +332,12 @@ fn template_injection() -> Result<()> {
             .run()?
     );
 
+    insta::assert_snapshot!(
+        zizmor()
+            .input(input_under_test("template-injection/pwsh-script.yml"))
+            .run()?
+    );
+
     Ok(())
 }